Add setDefaultLoanScheme rpc
